fix compile waring
git-svn-id: http://192.168.20.11/sequoiadb/trunk@31036 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/dms/dmsStorageLobData.cpp
@@ -518,7 +518,7 @@ namespace engine
       SDB_ASSERT( NULL != pData && offset >= 0, "invalid operation" ) ;
 
       INT32  pageID = -1 ;
-      if ( offset >= sizeof( _dmsStorageUnitHeader ) )
+      if ( offset >= (INT64)sizeof( _dmsStorageUnitHeader ) )
       {
          offset -= sizeof( _dmsStorageUnitHeader ) ;
          pageID = offset >> _logarithmic ;
@@ -670,7 +670,7 @@ namespace engine
 
       SINT64 readFromFile = 0 ;
       INT32  pageID = -1 ;
-      if ( offset >= sizeof( _dmsStorageUnitHeader ) )
+      if ( offset >= (INT64)sizeof( _dmsStorageUnitHeader ) )
       {
          offset -= sizeof( _dmsStorageUnitHeader ) ;
          pageID = offset >> _logarithmic ;
